Refactoring

+++ Assets/Scripts/Utils/Unity/PlayerPrefsEncryptionUtils.cs
@@ -19,7 +19,10 @@
         
         public static void DeleteKey(string key)
         {
-            ValidationUtils.ValidateStringNotNullOrEmpty(key, "key");
+            if (string.IsNullOrEmpty(key))
+            {
+                throw new ArgumentNullException("key");
+            }
 
             var encryptedKey = CipherUtility.Encrypt<RijndaelManaged>(key, SecuritySettings.PLAYERPREFS_PASSWORD, SecuritySettings.SALT);
             PlayerPrefs.DeleteKey(encryptedKey);
@@ -27,7 +30,10 @@
 
         public static bool HasKey(string key)
         {
-            ValidationUtils.ValidateStringNotNullOrEmpty(key, "key");
+            if (string.IsNullOrEmpty(key))
+            {
+                throw new ArgumentNullException("key");
+            }
 
             var encryptedKey = CipherUtility.Encrypt<RijndaelManaged>(key, SecuritySettings.PLAYERPREFS_PASSWORD, SecuritySettings.SALT);
             return PlayerPrefs.HasKey(encryptedKey);
@@ -35,8 +41,15 @@
 
         public static void SetString(string key, string value)
         {
-            ValidationUtils.ValidateStringNotNullOrEmpty(key, "key");
-            ValidationUtils.ValidateStringNotNullOrEmpty(key, "value");
+            if (string.IsNullOrEmpty(key))
+            {
+                throw new ArgumentNullException("key");
+            }
+
+            if (string.IsNullOrEmpty(value))
+            {
+                throw new ArgumentNullException("value");
+            }
 
             var encryptedKey = CipherUtility.Encrypt<RijndaelManaged>(key, SecuritySettings.PLAYERPREFS_PASSWORD, SecuritySettings.SALT);
             var encryptedValue = CipherUtility.Encrypt<RijndaelManaged>(value, SecuritySettings.PLAYERPREFS_PASSWORD, SecuritySettings.SALT);
@@ -45,7 +58,10 @@
 
         public static string GetString(string key)
         {
-            ValidationUtils.ValidateStringNotNullOrEmpty(key, "key");
+            if (string.IsNullOrEmpty(key))
+            {
+                throw new ArgumentNullException("key");
+            }
 
             var encryptedKey = CipherUtility.Encrypt<RijndaelManaged>(key, SecuritySettings.PLAYERPREFS_PASSWORD, SecuritySettings.SALT);
             var encryptedValue = PlayerPrefs.GetString(encryptedKey);
